requirements.txt
# ─── Telegram Bot ──────────────────────────────────────────────────────────────
python-telegram-bot>=20.7

# ─── HTTP / WebSocket ──────────────────────────────────────────────────────────
websockets>=12.0
aiohttp>=3.9.0

# ─── Price Data ────────────────────────────────────────────────────────────────
yfinance>=0.2.31

# ─── Config / Env ──────────────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ─── FastAPI Backend (backend/main.py) ─────────────────────────────────────────
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
